Prove pixel values cannot overflow during blending.
Enforce this at compile-time via StrictNumeric<uint8_t>. Consequently,
we can remove some DCHECK()s. In a subsequent CL, we can then argue
that an index into kColorSqrt[] must be within bounds.

+++ core/fxge/dib/cfx_scanlinecompositor.h
@@ -78,10 +78,10 @@ class CFX_ScanlineCompositor {
   FXDIB_Format m_SrcFormat;
   FXDIB_Format m_DestFormat;
   Palette m_SrcPalette;
-  int m_MaskAlpha;
-  int m_MaskRed;
-  int m_MaskGreen;
-  int m_MaskBlue;
+  uint8_t m_MaskAlpha;
+  uint8_t m_MaskRed;
+  uint8_t m_MaskGreen;
+  uint8_t m_MaskBlue;
   BlendMode m_BlendType = BlendMode::kNormal;
   bool m_bRgbByteOrder = false;
   bool m_bClip = false;
